+++ src/utils/scales/scale_continuous.test.ts
@@ -172,13 +172,29 @@ describe('Scale Continuous', () => {
     });
   });
 
+  describe('xScale values with minInterval and bandwidth', () => {
+    const domain = [7.053400039672852, 1070.1354763603908];
+
+    it('should return nice ticks when minInterval & bandwidth are 0', () => {
+      const scale = new ScaleContinuous(
+        {
+          type: ScaleType.Linear,
+          domain,
+          range: [0, 100],
+        },
+        { minInterval: 0, bandwidth: 0 },
+      );
+      expect(scale.ticks()).toEqual([100, 200, 300, 400, 500, 600, 700, 800, 900, 1000]);
+    });
+  });
+
   describe('time ticks', () => {
     const timezonesToTest = ['Asia/Tokyo', 'Europe/Berlin', 'UTC', 'America/New_York', 'America/Los_Angeles'];
 
     function getTicksForDomain(domainStart: number, domainEnd: number) {
       const scale = new ScaleContinuous(
         { type: ScaleType.Time, domain: [domainStart, domainEnd], range: [0, 100] },
-        { bandwidth: 0, minInterval: 0, timeZone: Settings.defaultZoneName },
+        { bandwidth: 0, minInterval: 0, timeZone: Settings.defaultZoneName, integersOnly: false },
       );
       return scale.tickValues;
     }
@@ -396,6 +412,22 @@ describe('Scale Continuous', () => {
           DateTime.fromISO('2019-03-01T00:00:00.000').toMillis(),
         ]);
       });
+    });
+  });
+  describe('ticks as integers or floats', () => {
+    const domain: Domain = [0, 7];
+    const minRange = 0;
+    const maxRange = 100;
+    let scale: ScaleContinuous;
+
+    beforeEach(() => {
+      scale = new ScaleContinuous({ type: ScaleType.Linear, domain, range: [minRange, maxRange] });
+    });
+    test('should return only integer ticks', () => {
+      expect(scale.getTicks(10, true)).toEqual([0, 1, 2, 3, 4, 5, 6, 7]);
+    });
+    test('should return normal ticks', () => {
+      expect(scale.getTicks(10, false)).toEqual([0, 0.5, 1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, 5, 5.5, 6, 6.5, 7]);
     });
   });
 });